+++ src/components/sheets/Sheets.jsx
@@ -27,7 +27,6 @@ const Sheets = () => {
       <input type="text" style={{fontFamily:"serif"}} value={title} onChange={(e)=>setTitle(e.target.value)} className=" title-input" placeholder={t("sheet.title.placeholder")}/>
       <ProfileCard />
       <Editor title={debouncedTitle} initialValue={initialValue}>
-        <Editor.Toolbar />
         <Editor.Input />
       </Editor>
     </div>
