+++ src/components/sheets/sheet-utils/Constant.js
@@ -1,3 +1,5 @@
+import { serialize } from "./serialize";
+
 const LIST_TYPES = ["ordered-list", "unordered-list"];
 const TEXT_ALIGN_TYPES = ["left", "center", "right", "justify"];
 const isAlignType = (format) => {
@@ -67,4 +69,33 @@ export {
   isListType,
   embedsRegex,
   removeFootnotes,
+};
+
+export const createPayload = (value, title, is_published = false) => {
+  const source = value.map((node, i) => {
+    if (["image", "audio", "video"].includes(node.type)) {
+      return {
+        position: i,
+        type: node.type,
+        content: node.src,
+      };
+    }
+    if (node.type === "pecha") {
+      return {
+        position: i,
+        type: "segment",
+        content: node.src,
+      };
+    }
+    return {
+      position: i,
+      type: "content",
+      content: serialize(node),
+    };
+  });
+  return {
+    title,
+    source,
+    is_published,
+  };
 };
